requirements.txt
google-generativeai
aiohttp
python-dotenv
sentence-transformers
torch